requirements.txt
numpy
tqdm
openai-whisper
python-docx
